cookie change interceptor
The problem with SessionChangeInterceptor is that after session is created
`;jsessionid=` fragment is added to all generated URLs. This is a backup
solution when user does not support session. This behaviour is a security
risk. What is more we are not using session in rimrock at all. That is why
the way how language preffered by the user was changed into storing this
information into `rimrock-lang` cookie value. It is possible thanks to
`CookieChangeInterceptor`.